Initial Commit
I need to clean up the code a bit and restructure the Map/MapFrag
classes.  The method that determines which iso cell is touched needs to
be fixed so it works in both verticle and horizontal.  As it stands, it
probably only works vertically on my phone.

Future versions will support adding objects into the map.