Activity doesn't Reset | Context dereference
The activity handles orientation changes, screen size changes (needed
for orientation changes since API 13), and keyboard hidden (which
shouldn't be an issue for quite a while).

The context variable needed to initialize renderer, Iso, Map,
AppControl, and Obj was removed and replaced with a static variable in
HexMapEditorActivity and a static method in the same class that returned
it.